requirements.txt
# local package
# Runnable via Python 3.6
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas
tweepy
twitter
PyYAML
python-twitter